Modif unsign type for byte operation

+++ commands/ldi.c
@@ -24,8 +24,9 @@ t_b_command     *ldi(char **command, t_champ *champ)
     champ->command_size += (byte_command->arg1->size + 1);
     byte_command->arg2 = get_arg(command[1], LDI_T_DIR_SIZE, champ->labels_vec);
     champ->command_size += (byte_command->arg2->size + 1);
-    byte_command->arg_type_code = (byte_command->arg1->type << 6) + (byte_command->arg2->type << 4) +
-            (byte_command->arg3->type << 2);
+    byte_command->arg_type_code = (byte_command->arg1->type << (unsigned int)6) +
+                                  (byte_command->arg2->type << (unsigned int)4) +
+                                  (byte_command->arg3->type << (unsigned int)2);
 
     ft_ptr_vec_pushback(champ->command_vec, byte_command);
     return (byte_command);
